Improved database related stuff)
