+++ src/main/scala/db/RecordMapper.scala
@@ -1,0 +1,9 @@
+package db {
+
+  trait RecordMapper[Record, T] {
+    def fromRecord(record: Record): T
+
+    def toRecord(model: T): Record
+  }
+
+}
