+++ src/main/scala/user/UserRoute.scala
@@ -6,8 +6,6 @@ import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
 import spray.json.DefaultJsonProtocol._
 import spray.json._
-import user.dao.UserDao
-import user.model.User
 
 class UserRoute {
   implicit val userFormat: RootJsonFormat[User] = jsonFormat3(User.apply)
@@ -19,7 +17,7 @@ class UserRoute {
   }
 
   def getSelf: Route = {
-    new UserDao().getUser.map {
+    UserDao.getUser.map {
       case Left(error) => complete(StatusCodes.BadRequest, error.reason)
       case Right(user) => complete(user)
     }.getOrElse(complete(StatusCodes.NotFound))
